regulator: Add notification support
For regulator, it is desirable to have a generic notification mechanism
available. This mechanism can be used to notify userspace of the driver
events on status & state change.

This is useful for regulator consumer control from userspace to track
its state & status change.

Notification will not happend too often as they happen only when there
are critical events. Thus would not cause performance impact.

Signed-off-by: Patrick Rudolph <patrick.rudolph@9elements.com>
Signed-off-by: Naresh Solanki <Naresh.Solanki@9elements.com>