hwmon: (max6639) Add option to disable !THERM assertion
Add workaround for undocumented behaviour.

According to the datasheet [1] on the first and 4th page the !THERM pin
is an output, however in other places it becomes clear that is actually
a bi-directional pin.
What's not mentioned at all is that it needs an external pull-up resistor
when not being used, to prevent it from falsely detecting a thermal alert.

On mainboards where the pin is left floating this causes the PWM to be
forced to 100%, making fan regulation impossible.

Add support to disable !THERM assertion via the optional
"maxim,disable_therm_full_speed" property to allow proper PWM fan regulation.

1: https://datasheets.maximintegrated.com/en/ds/MAX6639-MAX6639F.pdf

Signed-off-by: Patrick Rudolph <patrick.rudolph@9elements.com>
Signed-off-by: Marcello Sylvester Bauer <sylv@sylv.io>