aspeed-bmc-ibm-genesis3: Drop constrains
The max597x is now handled as "switch", that means it always provides
the same voltage as the parent regulator. As it cannot regulate the output
in any way, drop the constrains, which hadn't any effect anyways.

The regulator core will provide to correct voltage readings automatically
now that is a switch.

Signed-off-by: Patrick Rudolph <patrick.rudolph@9elements.com>