aspeed-espi-ctrl: Get rid of ioctls and use MTD subsystem
The eSPI specification allows to share the flash in a transparent
fashion on either side of the wire called MAFS (Master attached storage)
and SAFS (Slave attached storage). However it doesn't specify the
maximum size of the flash, requiring to hardcode the size.
As usually the flash has multiple partitions, of which some are protected
for security reasons, the parition layout needs to be hardcoded as well.

The AST2500 supports both using software sequencing, while the AST2600
also supports SAFS in HW sequencing mode.

This driver implements the MAFS part on the BMC side and exposes the
flash on "the other side" as MTD device.

TODO: Implement SAFS software mode using the spi-nor driver.

Tested: Could R/W/E the whole SPI flash connected to Intel Lewisburg PCH.

Signed-off-by: Patrick Rudolph <patrick.rudolph@9elements.com>