pmbus: Introduce PMBUS_HAVE_PGODD
The PMBUS specification [1] allows for an optional POWER_GOOD status
bit, that can be found on many devices. Add a new functionality bit
in order to mark such devices supporting the POWER_GOOD bit.

This bit can be used to determine the current regulator status and
will be used in the following patch series.

1: https://pmbus.org/specification-archives/

Signed-off-by: Patrick Rudolph <patrick.rudolph@9elements.com>