[NOT FOR UPSTREAM][FIXME]pmbus_core: Watch PMBUS status register until regulator is actually turned on
The introduced PGOOD status bit allows to wait in the regulator core for the
device to be power on. For this the poll_enabled_time must be set.

TODO: Figure out what a good default is.
TODO: Move poll_enabled_time to regulators. Figure out how to check for PMBUS_HAVE_PGOOD.

Signed-off-by: Patrick Rudolph <patrick.rudolph@9elements.com>