hwmon/pmbus: mp2975: Simplify VOUT code
In order to upstream MP2973/MP2971 simplify the code by removing support for
various VOUT formats. The MP2973 and MP2971 supports all PMBUS supported formats
for VOUT, while the MP2975 only support DIRECT and VID for VOUT.

In DIRECT mode all chips report the voltage in 1mV/LSB.

Configure the chip to use DIRECT format for VOUT and drop the code conversion
code for other formats. The to be added chips MP2973/MP2971 will be configured
to also report VOUT in DIRECT format.

The maximum voltage that can be reported in DIRECT format is 4096mV. This is
sufficient as the maximum output voltage for VR12/VR13 is 3040 mV.

Signed-off-by: Patrick Rudolph <patrick.rudolph@9elements.com>