hwmon/pmbus: mp2975: Disable unsupported features
On MP2973 and MP2971 the Vref cannot be read and thus most of the
calculations won't work. Guard features only working on MP2975 and
disable reading of Vref, Vref offset, Vout max, Vout OV scale,
VRM version and current scale registers.

The following features are not exposed on MP2973/MP2971:
 - Per phase current reporting
 - OV_FAULT_LIMIT reporting
 - UV_FAULT_LIMIT reporting

Signed-off-by: Patrick Rudolph <patrick.rudolph@9elements.com>