thunderbolt: Add Dell WD19TB reset workaround
The Dell WD19TB dock causes problems on Alder Lake when the DP tunnel
is created by the boot firmware, e.g. a QEMU Windows guest with Intel
GPU passthrough is sometimes unable to activate an external monitor
without a Thunderbolt hotplug. To mitigate this problem, do a switch
reset if it is already enabled, and abort the switch initialization.
This should cause a new hotplug event to initialize that device
correctly.

Tested on:
- Lenovo Thinkpad T14 G3
- Dell Latitude 7430 2-in-1

Signed-off-by: Marcello Sylvester Bauer <sylv@sylv.io>