hwmon: (pmbus/tda38640) Add workaround for bug in SVID mode
The TDA38640 supports two operating modes to set the output voltage:
- PMBUS
- SVID

Due to an undocumented bug the regulator cannot be enabled using BIT7
of OPERATION(01h) register when in SVID mode. It works when in PMBUS
mode. In SVID mode the regulator only cares about the ENABLE pin.

Add a workaround that needs the ENABLE pin to be left floating or
tied to a fixed level. The DT must contain the newly introduced
property 'infineon,en-pin-fixed-level' to enable this workaround.

The workaround will map the PB_OPERATION_CONTROL_ON bit to the
PB_ON_OFF_CONFIG_EN_PIN_REQ bit of the ON_OFF_CONFIG register.
In combination with the fixed level on the ENABLE pin the regulator
can be controlled as expected.

Signed-off-by: Patrick Rudolph <patrick.rudolph@9elements.com>