i2c: muxes: Enable features on MAX7357/MAX7358
Detect that max7357 is being used and run custom init sequence.
Enable additional features based on DT settings and unconditionally
release the shared interrupt pin after 1.6 seconds and allow to use
it as reset.

The init sequence is not run for max7358 that needs to be unlocked
first, but that would need the unimplemented function
i2c_probe_func_quick_write().

Signed-off-by: Patrick Rudolph <patrick.rudolph@9elements.com>