i2c: muxes: Enable features on MAX7357
Detect that max7357 is being used and run custom init sequence.

By default MAX7357 disconnects all channels on a bus lock-up and
signals	this condition to the bus master using an interrupt.

Disable the interrupt as it's not useful within the kernel and
it might conflict with the reset functionality that shares the same
pin.

Use the introduced 'maxim,bus-lockup-fix' property to enable
faulty channel isolation and flush-out sequence generation.

Signed-off-by: Patrick Rudolph <patrick.rudolph@9elements.com>
Signed-off-by: Naresh Solanki <Naresh.Solanki@9elements.com>