pinctrl: cy8c95x0: Fix non push-pull pin modes
When configuring a pin in open-drain or open-source mode the kernel
will switch the direction when toggling the pin in order to read the
current line value. This caused the push-pull mode to get activated
when driving the line and enabled high-z mode when reading the line.
Those changes to the drive-mode are unintended and accidentally disabled
the previously configured drive-mode or activated bias.

Rework the bias/drive-mode logic to fix this issue and properly
support open-drain/open-source with/without configured bias.

1. By default the push-pull mode will be used:
   In this mode the 'output enable' will set drive mode to push-pull
   and 'input enable' will set drive mode high-z.

2. By setting a bias or drive-mode different than push-pull:
   It disables the behaviour mentioned in 1. Setting the 'output enable'
   or 'input enable' will no longer touch drive-mode bits, keeping the
   bias or drive-mode currently active.

3. By setting a bias-disable or drive mode push-pull:
   The pin will be reset back to push-pull mode and behave as
   described in 1.

Signed-off-by: Patrick Rudolph <patrick.rudolph@9elements.com>