pinctrl: cy8c95x0: Fix non push-pull pin modes
When configuring a pin in open-drain or open-source mode the kernel
will switch the direction when toggling the pin. This caused the
push-pull mode to get activated, which in turn let to high-z when
switched to input. Those changes to the drive-mode are unintended
and accidentally disabled the previously activated bias.

Rework the bias/drive-mode logic to fix this issue and properly
support open-drain/open-source with/without configured bias.

By default the push-pull mode will be used. In this mode the output
enable will set drive mode to push-pull and input enable will set
drive mode high-z.
By setting a bias or drive mode different than push-pull, it will
disable the above mentioned behaviour. Setting the output enable
or input enable will no longer touch drive-mode, keeping the bias
or drive mode currently active.
By setting a bias-disable or drive mode push-pull the pin will be
reset back to push-pull.

Signed-off-by: Patrick Rudolph <patrick.rudolph@9elements.com>