hwmon: pmbus: Implement generic bus access delay
Some drivers, like the max15301 or zl6100, are delaying SMBus communication
to prevent transmission errors. As this is necessary on additional
PMBus compatible devices, implement a generic delay mechanism in the
pmbus core.

Introduces two delay settings in the pmbus_driver_info struct, one applies to
every SMBUS transaction and the other is for write transaction only. Once set
by the driver the SMBUS traffic using the generic pmbus access helpers is
delayed when necessary.

This allows to drop the custom delay code from the drivers and easily
introduce this feature in additional pmbus drivers.

Signed-off-by: Patrick Rudolph <patrick.rudolph@9elements.com>