hwmon: pmbus: pli12096bc: Add write delay
Tests on PLI12096bc showed that sometimes a small delay is necessary
after a write operation before a new operation can be processed.
If not respected the device will likely NACK the data phase of
the SMBus transaction.

Further tests showed that 300 microseconds, as used previously for
the CLEAR_FAULTS workaround, is sufficient.

Test: Done 100000 writes and seen no read failures afterwards.

Signed-off-by: Patrick Rudolph <patrick.rudolph@9elements.com>