regulator/userspace-consumer: Keep track of every regulator
When a regulator fails to enable or disable it's impossible to
tell from the return code if the refcount was modified and another
attempt can be safely made without violating the balanced
enabled/disable call requirement.

In order to prevent any refcount issue and to keep the internal state
always in sync with the real regulator state use the event callback
to keep track of every single regulator.

This also works fine on errors and always keeps the internal state
in sync with the regulators state.

Signed-off-by: Patrick Rudolph <patrick.rudolph@9elements.com>