regulator: userspace-consumer: Fix refcount
When the referenced regulator is already enabled by some
other consumer the internal state is set to enabled, but the
regulator's refcount isn't increased by calling regulator_enable.
This can lead to a condition where a regulator_disable will fail
due to refcount underflow.

Fix this corner case by increasing refcount when the internal
state is set to enabled.

Signed-off-by: Patrick Rudolph <patrick.rudolph@9elements.com>