regulator: userspace-consumer: Fix refcount
At the time of probing when the referenced regulator was already enabled
by some other consumer or is enabled by default the internal state is set
to enabled, but the regulator's refcount isn't increased by calling
regulator_bulk_enable(). This can lead to a refcount underflow when the
user later disables the regulator using the sysfs.

Fix this corner case by increasing refcount when the internal state
is set to enabled at probe.

Signed-off-by: Patrick Rudolph <patrick.rudolph@9elements.com>