ARM: dts: aspeed: yosemite4: Add i2c-mux for four NICs
Add i2c-mux on Spider board for four NICs and add the temperature sensor
and EEPROM for the NICs.

Also remove the mctp-controller property on I2C bus 15 because we need
to add the property on the I2C mux to each NIC so that the MCTP driver
will ensure that each port is configured properly before communicating
with the NICs.

Signed-off-by: Ricky CX Wu <ricky.cx.wu.wiwynn@gmail.com>
Signed-off-by: Delphine CC Chiu <Delphine_CC_Chiu@wiwynn.com>
Message-ID: <20241003074251.3818101-3-Delphine_CC_Chiu@wiwynn.com>
Signed-off-by: Andrew Jeffery <andrew@codeconstruct.com.au>